update: all variation done with requirments
