+++ src/app/admin/arm-types/[id]/edit/page.tsx
@@ -112,9 +112,6 @@ const EditArmTypePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Arm Type
-        </Typography>
 
         {/* Alerts */}
         {error && (
