+++ src/app/admin/arm-types/page.tsx
@@ -152,9 +152,7 @@ const ArmTypesPage = () => {
           gap: { xs: 2, sm: 0 }
         }}>
           <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2, flex: 1 }}>
-            <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-              Arm Types
-            </Typography>
+          
             {/* Search Bar positioned at top-left */}
             <TextField
               placeholder="Search arm types..."
