+++ src/app/admin/color-vendors/[id]/edit/page.tsx
@@ -135,9 +135,7 @@ const EditColorVendorPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Color Vendor
-        </Typography>
+      
 
         {/* Alerts */}
         {error && (
