--- src/app/admin/color-vendors/[id]/page.tsx
@@ -1,224 +1,0 @@
-'use client';
-
-import React, { useState, useEffect } from 'react';
-import { useRouter, useParams } from 'next/navigation';
-import {
-  Box,
-  Typography,
-  Button,
-  Paper,
-  Alert,
-  Chip,
-  Stack,
-  CircularProgress,
-  Divider,
-} from '@mui/material';
-import { ArrowBack as ArrowBackIcon, Edit as EditIcon } from '@mui/icons-material';
-import AdminLayout from '@/components/AdminLayout';
-import { apiService } from '@/utils/api';
-
-interface ColorVendor {
-  id: number;
-  name: string;
-  description: string;
-  is_active: boolean;
-  sort_order: number;
-  created_at: string;
-  updated_at: string;
-}
-
-const ColorVendorDetailPage = () => {
-  const router = useRouter();
-  const params = useParams();
-  const id = params.id as string;
-  
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [colorVendor, setColorVendor] = useState<ColorVendor | null>(null);
-
-  useEffect(() => {
-    loadColorVendor();
-  }, [id]);
-
-  const loadColorVendor = async () => {
-    try {
-      setLoading(true);
-      setError(null);
-      
-      const data = await apiService.getColorVendor(parseInt(id));
-      setColorVendor(data);
-    } catch (err: any) {
-      setError(err.message || 'Failed to load color vendor');
-      console.error('Error loading color vendor:', err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleEdit = () => {
-    router.push(`/admin/variations/color-vendors/${id}/edit`);
-  };
-
-  const handleBack = () => {
-    router.push('/admin/variations/color-vendors');
-  };
-
-  const formatDate = (dateString: string) => {
-    return new Date(dateString).toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  };
-
-  if (loading) {
-    return (
-      <AdminLayout title="Color Vendor Details">
-        <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', minHeight: '400px' }}>
-          <CircularProgress />
-        </Box>
-      </AdminLayout>
-    );
-  }
-
-  if (error || !colorVendor) {
-    return (
-      <AdminLayout title="Color Vendor Details">
-        <Box>
-          <Box sx={{ mb: 3, display: 'flex', alignItems: 'center', gap: 2 }}>
-            <Button
-              startIcon={<ArrowBackIcon />}
-              onClick={handleBack}
-              sx={{ color: 'text.secondary' }}
-            >
-              Back to Color Vendors
-            </Button>
-          </Box>
-
-          <Alert severity="error" sx={{ mb: 3 }}>
-            {error || 'Color vendor not found'}
-          </Alert>
-        </Box>
-      </AdminLayout>
-    );
-  }
-
-  return (
-    <AdminLayout title="Color Vendor Details">
-      <Box>
-        {/* Header */}
-        <Box sx={{ mb: 3, display: 'flex', alignItems: 'center', gap: 2 }}>
-          <Button
-            startIcon={<ArrowBackIcon />}
-            onClick={handleBack}
-            sx={{ color: 'text.secondary' }}
-          >
-            Back to Color Vendors
-          </Button>
-        </Box>
-
-        <Box sx={{ mb: 3, display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Color Vendor Details
-          </Typography>
-          <Button
-            variant="contained"
-            startIcon={<EditIcon />}
-            onClick={handleEdit}
-          >
-            Edit Color Vendor
-          </Button>
-        </Box>
-
-        {/* Content */}
-        <Paper sx={{ p: 4 }}>
-          <Box sx={{ display: 'flex', flexDirection: { xs: 'column', md: 'row' }, gap: 4 }}>
-            <Box sx={{ flex: 1 }}>
-              <Stack spacing={3}>
-                <Box>
-                  <Typography variant="h6" gutterBottom sx={{ fontWeight: 600 }}>
-                    Name
-                  </Typography>
-                  <Typography variant="body1">
-                    {colorVendor.name}
-                  </Typography>
-                </Box>
-
-                <Box>
-                  <Typography variant="h6" gutterBottom sx={{ fontWeight: 600 }}>
-                    Description
-                  </Typography>
-                  <Typography variant="body1" color="text.secondary">
-                    {colorVendor.description || 'No description available'}
-                  </Typography>
-                </Box>
-
-                <Box>
-                  <Typography variant="h6" gutterBottom sx={{ fontWeight: 600 }}>
-                    Status
-                  </Typography>
-                  <Chip
-                    label={colorVendor.is_active ? 'Active' : 'Inactive'}
-                    color={colorVendor.is_active ? 'success' : 'default'}
-                    size="small"
-                  />
-                </Box>
-
-                <Box>
-                  <Typography variant="h6" gutterBottom sx={{ fontWeight: 600 }}>
-                    Sort Order
-                  </Typography>
-                  <Typography variant="body1">
-                    {colorVendor.sort_order}
-                  </Typography>
-                </Box>
-              </Stack>
-            </Box>
-
-            <Box sx={{ width: { xs: '100%', md: 300 } }}>
-              <Paper variant="outlined" sx={{ p: 3 }}>
-                <Typography variant="h6" gutterBottom sx={{ fontWeight: 600 }}>
-                  Information
-                </Typography>
-                <Divider sx={{ mb: 2 }} />
-                
-                <Stack spacing={2}>
-                  <Box>
-                    <Typography variant="body2" color="text.secondary">
-                      ID
-                    </Typography>
-                    <Typography variant="body1" sx={{ fontWeight: 500 }}>
-                      {colorVendor.id}
-                    </Typography>
-                  </Box>
-
-                  <Box>
-                    <Typography variant="body2" color="text.secondary">
-                      Created
-                    </Typography>
-                    <Typography variant="body1" sx={{ fontWeight: 500 }}>
-                      {formatDate(colorVendor.created_at)}
-                    </Typography>
-                  </Box>
-
-                  <Box>
-                    <Typography variant="body2" color="text.secondary">
-                      Last Updated
-                    </Typography>
-                    <Typography variant="body1" sx={{ fontWeight: 500 }}>
-                      {formatDate(colorVendor.updated_at)}
-                    </Typography>
-                  </Box>
-                </Stack>
-              </Paper>
-            </Box>
-          </Box>
-        </Paper>
-      </Box>
-    </AdminLayout>
-  );
-};
-
-export default ColorVendorDetailPage;
