+++ src/app/admin/color-vendors/create/page.tsx
@@ -93,9 +93,7 @@ const CreateColorVendorPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Color Vendor
-        </Typography>
+    
 
         {/* Alerts */}
         {error && (
