+++ src/app/admin/colors/page.tsx
@@ -207,9 +207,7 @@ const ColorsPage = () => {
           gap: { xs: 2, sm: 0 }
         }}>
           <Box sx={{ display: 'flex', flexDirection: 'column', gap: 2, flex: 1 }}>
-            <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-              Colors
-            </Typography>
+       
             {/* Search Bar positioned at top-left */}
             <TextField
               placeholder="Search colors..."
