+++ src/app/admin/feature-options/[id]/page.tsx
@@ -117,9 +117,7 @@ const FeatureOptionDetailPage = () => {
         </Box>
 
         <Box sx={{ mb: 3, display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Feature Option Details
-          </Typography>
+        
           <Button
             variant="contained"
             startIcon={<EditIcon />}
