+++ src/app/admin/feature-options/create/page.tsx
@@ -80,10 +80,7 @@ const CreateFeatureOptionPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Feature Option
-        </Typography>
-
+ 
         {/* Alerts */}
         {error && (
           <Alert severity="error" sx={{ mb: 3 }}>
