+++ src/app/admin/heat-options/[id]/edit/page.tsx
@@ -192,9 +192,7 @@ const EditHeatOptionPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Heat Option
-        </Typography>
+    
 
         {/* Alerts */}
         {error && (
