+++ src/app/admin/item-types/[id]/edit/page.tsx
@@ -116,9 +116,7 @@ const EditItemTypePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Item Type
-        </Typography>
+      
 
         {/* Alerts */}
         {error && (
