+++ src/app/admin/item-types/[id]/page.tsx
@@ -117,9 +117,7 @@ const ItemTypeDetailPage = () => {
         </Box>
 
         <Box sx={{ mb: 3, display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Item Type Details
-          </Typography>
+        
           <Button
             variant="contained"
             startIcon={<EditIcon />}
