+++ src/app/admin/lumbar-types/[id]/edit/page.tsx
@@ -273,9 +273,7 @@ const EditLumbarTypePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Lumbar Type
-        </Typography>
+      
 
         {/* Alerts */}
         {error && (
