+++ src/app/admin/lumbar-types/create/page.tsx
@@ -214,7 +214,7 @@ const CreateLumbarTypePage = () => {
   };
 
   return (
-    <AdminLayout title="Create Lumbar Type">
+    <AdminLayout title="Create New Lumbar Type">
       <Box>
         {/* Header */}
         <Box sx={{ mb: 3, display: 'flex', alignItems: 'center', gap: 2 }}>
@@ -227,9 +227,7 @@ const CreateLumbarTypePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Lumbar Type
-        </Typography>
+      
 
         {/* Alerts */}
         {error && (
