+++ src/app/admin/lumbar-types/page.tsx
@@ -175,13 +175,11 @@ const LumbarTypesPage = () => {
           mb: 3, 
           display: 'flex', 
           flexDirection: { xs: 'column', sm: 'row' },
-          justifyContent: 'space-between', 
+          justifyContent: 'flex-end', 
           alignItems: { xs: 'stretch', sm: 'center' },
           gap: { xs: 2, sm: 0 }
         }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Lumbar Types
-          </Typography>
+       
           <Button
             variant="contained"
             startIcon={<AddIcon />}
@@ -195,7 +193,7 @@ const LumbarTypesPage = () => {
               }
             }}
           >
-            Add Lumbar Type
+            Add
           </Button>
         </Box>
