+++ src/app/admin/material-types/page.tsx
@@ -180,13 +180,11 @@ const MaterialTypesPage = () => {
           mb: 3, 
           display: 'flex', 
           flexDirection: { xs: 'column', sm: 'row' },
-          justifyContent: 'space-between', 
+          justifyContent: 'flex-end', 
           alignItems: { xs: 'stretch', sm: 'center' },
           gap: { xs: 2, sm: 0 }
         }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Material Types
-          </Typography>
+        
           <Button
             variant="contained"
             startIcon={<AddIcon />}
@@ -200,7 +198,7 @@ const MaterialTypesPage = () => {
               }
             }}
           >
-            Add Material Type
+            Add
           </Button>
         </Box>
