+++ src/app/admin/recline-types/[id]/edit/page.tsx
@@ -177,9 +177,7 @@ const EditReclineTypePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Recline Type
-        </Typography>
+      
 
         {/* Alerts */}
         {error && (
