+++ src/app/admin/recline-types/page.tsx
@@ -179,13 +179,11 @@ const ReclineTypesPage = () => {
           mb: 3, 
           display: 'flex', 
           flexDirection: { xs: 'column', sm: 'row' },
-          justifyContent: 'space-between', 
+          justifyContent: 'flex-end', 
           alignItems: { xs: 'stretch', sm: 'center' },
           gap: { xs: 2, sm: 0 }
         }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Recline Types
-          </Typography>
+       
           <Button
             variant="contained"
             startIcon={<AddIcon />}
@@ -199,7 +197,7 @@ const ReclineTypesPage = () => {
               }
             }}
           >
-            Add Recline Type
+            Add
           </Button>
         </Box>
