+++ src/app/admin/seat-pricing/[id]/edit/page.tsx
@@ -124,9 +124,7 @@ const EditSeatPricingPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Seat Pricing
-        </Typography>
+       
 
         {/* Alerts */}
         {error && (
