+++ src/app/admin/seat-pricing/create/page.tsx
@@ -88,9 +88,7 @@ const CreateSeatPricingPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Seat Pricing
-        </Typography>
+      
 
         {/* Alerts */}
         {error && (
