+++ src/app/admin/seat-stitch-patterns/[id]/edit/page.tsx
@@ -193,9 +193,7 @@ const EditSeatStitchPatternPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Edit Seat Stitch Pattern
-        </Typography>
+       
 
         {/* Alerts */}
         {error && (
