+++ src/app/admin/seat-stitch-patterns/create/page.tsx
@@ -152,9 +152,7 @@ const CreateSeatStitchPatternPage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Seat Stitch Pattern
-        </Typography>
+    
 
         {/* Alerts */}
         {error && (
