+++ src/app/admin/seat-styles/create/page.tsx
@@ -77,9 +77,7 @@ const CreateSeatStylePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Seat Style
-        </Typography>
+     
 
         {/* Alerts */}
         {error && (
