+++ src/app/admin/seat-types/create/page.tsx
@@ -77,9 +77,7 @@ const CreateSeatTypePage = () => {
           </Button>
         </Box>
 
-        <Typography variant="h4" component="h1" gutterBottom sx={{ fontWeight: 600 }}>
-          Create New Seat Type
-        </Typography>
+     
 
         {/* Alerts */}
         {error && (
