+++ src/app/admin/seat-types/page.tsx
@@ -152,13 +152,11 @@ const SeatTypesPage = () => {
           mb: 3, 
           display: 'flex', 
           flexDirection: { xs: 'column', sm: 'row' },
-          justifyContent: 'space-between', 
+          justifyContent: 'flex-end', 
           alignItems: { xs: 'stretch', sm: 'center' },
           gap: { xs: 2, sm: 0 }
         }}>
-          <Typography variant="h4" component="h1" sx={{ fontWeight: 600 }}>
-            Seat Types
-          </Typography>
+        
           <Button
             variant="contained"
             startIcon={<AddIcon />}
@@ -172,7 +170,7 @@ const SeatTypesPage = () => {
               }
             }}
           >
-            Add Seat Type
+            Add
           </Button>
         </Box>
